Refactor: extract logic to create test queues
This is going to be used for sessions+party test as well
